Fix issue #413 Removing blog directory makes blogs unremovable
- If the option "delete blog index only" wasn't selected and the user already removed the blog's folder manually, an error happened on "remove blog".

+++ src/TumblThree/TumblThree.Applications/Controllers/ManagerController.cs
@@ -735,7 +735,8 @@ namespace TumblThree.Applications.Controllers
                     try
                     {
                         string blogPath = blog.DownloadLocation();
-                        Directory.Delete(blogPath, true);
+                        if (Directory.Exists(blogPath))
+                            Directory.Delete(blogPath, true);
                     }
                     catch (Exception ex)
                     {
